ExternalTools: create one central KSharedConfig object and use it everywhere
The config file was saved in ~/.local/share/application/externaltools, which
is a slightly weird location on Linux, instead use GenericConfigLocation, so
it's saved in ~/.config and the equivalent path in other OS's.

Also rename the file to kateexternaltoolsrc, this is more consisten with how
other config files are named.