Add a frontend for compiler-explorer
This change introduces a new plugin: Compiler Explorer. The main purpose
of this plugin is to allow you to view the generated code of any file
in your project easily. You can do that in Compiler-Explorer but it
takes a lot more effort and you have to copy over flags etc.

Besides the plugin itself, there is a major change packed in i.e., the
ability to allow *any* QWidget in Kate's tabs.